WIP: Add noscript functionality
This is WIP, but fairly functional: Add functionality to only allow
loading javascript from a specific set of domains, and expose it using
the "javascript-whitelist" variable.

Note that this currently does *not* disable in-line javascript, and it
is probably quite slow for large whitelists. The API is probably also
far from ready.

You should honestly only use this if you *really* don't want to
execute javascript.